requirements.txt
alembic==1.18.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Flask==3.1.2
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.9
pycparser==2.23
python-docx==1.2.0
python-http-client==3.3.7
sendgrid==6.12.5
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.5